requirements.txt
numpy
pandas
scikit-learn
scipy
matplotlib
pytest
nltk
umap-learn
hdbscan
